requirements.txt
ultralytics
numpy
pandas
supervision
matplotlib
opencv-python
roboflow
customtkinter
Pillow